move course enrollment into models folder
